refactor(ui): remove dedicated/normal mode distinction on select
Consolidate cloud practitioner start into the main select page and
remove separate mode pages to reduce learner confusion.

Made-with: Cursor
