--- app/select/cloud-practitioner/page.tsx
@@ -1,11 +1,0 @@
-import { requireUser } from "@/lib/auth/guards";
-
-import { CloudPractitionerSelectForm } from "./select-form";
-
-const CloudPractitionerSelectPage = async (): Promise<React.ReactElement> => {
-  await requireUser();
-
-  return <CloudPractitionerSelectForm />;
-};
-
-export default CloudPractitionerSelectPage;
