+++ app/select/select-form.tsx
@@ -9,6 +9,17 @@ type CategoryInfo = {
   count: number;
 };
 
+const CLOUD_PRACTITIONER_CATEGORIES = [
+  "VPC",
+  "EC2",
+  "S3",
+  "IAM",
+  "CloudWatch",
+  "CloudTrail",
+  "RDS",
+  "Lambda",
+] as const;
+
 export const SelectForm = () => {
   const router = useRouter();
   const [categories, setCategories] = useState<CategoryInfo[]>([]);
@@ -64,6 +75,37 @@ export const SelectForm = () => {
 
   const deselectAllCategories = (): void => {
     setSelectedCategories([]);
+  };
+
+  const handleStartCloudPractitioner = async (): Promise<void> => {
+    setError("");
+    setIsSubmitting(true);
+
+    try {
+      const response = await fetch("/api/attempts/create", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          preset: "cloud-practitioner",
+          categories: CLOUD_PRACTITIONER_CATEGORIES,
+          levels: [1, 2, 3],
+          count: 30,
+        }),
+      });
+
+      if (!response.ok) {
+        const data = (await response.json()) as { message?: string };
+        setError(data.message ?? "エラーが発生しました");
+        return;
+      }
+
+      const data = (await response.json()) as { attemptId: string };
+      router.push(`/quiz/${data.attemptId}`);
+    } catch {
+      setError("通信に失敗しました。ネットワーク接続を確認してください");
+    } finally {
+      setIsSubmitting(false);
+    }
   };
 
   const handleSubmit = async (
@@ -133,18 +175,19 @@ export const SelectForm = () => {
       <h1 className="mb-6 text-2xl font-semibold">問題条件を選択</h1>
 
       <section className="mb-6 rounded-xl border border-brand-200/60 bg-brand-50/70 p-4 dark:border-brand-500/40 dark:bg-brand-900/20">
-        <h2 className="text-sm font-semibold text-brand-700 dark:text-brand-200">
-          学習モード
-        </h2>
+        <h2 className="text-sm font-semibold text-brand-700 dark:text-brand-200">おすすめ設定</h2>
         <p className="mt-1 text-xs text-neutral-700 dark:text-neutral-300">
-          AWS Certified Cloud Practitioner 向けの固定プリセットで学習する場合は、専用モードから開始できます。
+          AWS Certified Cloud Practitioner 向けの推奨設定（カテゴリ固定・Lv.1〜3混在・30問）です。
         </p>
         <button
           type="button"
-          onClick={() => router.push("/select/cloud-practitioner")}
+          onClick={() => {
+            void handleStartCloudPractitioner();
+          }}
+          disabled={isSubmitting}
           className="mt-3 rounded-lg border border-brand-400 bg-white px-3 py-2 text-xs font-medium text-brand-700 transition hover:bg-brand-100 dark:border-brand-300 dark:bg-black/30 dark:text-brand-200 dark:hover:bg-brand-900/40"
         >
-          Cloud Practitioner専用モードへ
+          Cloud Practitioner推奨設定で開始
         </button>
       </section>
